fix(logs): harden GCP/Azure/CF/DO collectors (review findings) (#220)
- GCP/Azure tail polls were re-scanning the whole backfill window every tick
  (gcp 30d, azure 24h) AND could skip a burst >limit between ticks. Poll
  incrementally now: GCP filters timestamp>lastSeen (asc, 1000); Azure uses
  --start-time lastSeen. Track max-seen ts; bounded, no skips, no rescan.
- Azure: dedup by unique eventDataId/correlationId (was empty stream → distinct
  same-ms events dropped); sort results newest-first instead of trusting CLI
  order for 'last N'.
- Cloudflare chat fed the model zero logs (Query is a no-op for live-only):
  chat now samples the live tail for a bounded 8s window. Timestamps parsed as
  float64 so a non-integer JSON value no longer silently drops the event.
- DigitalOcean: parse a leading RFC3339 timestamp when present (ordering +
  time-filtering) instead of always stamping now.
- Severity: GCP DEFAULT → unknown (not debug); Cloudflare 'log' → info.
- refDedup prunes by max-seen ts, not the just-added (possibly older) entry.
- chat prompt: print 'last N entries' instead of a zero-value window-from.
Verified GCP live (query + incremental tail). Adds severity edge-case test.

+++ internal/logs/cloudflare.go
@@ -38,18 +38,18 @@ func (c *cfCollector) Query(ctx context.Context, opts Options, emit Emit) error 
 
 // cfTailEvent is one `wrangler tail --format json` record (a Worker invocation).
 type cfTailEvent struct {
-	ScriptName     string `json:"scriptName"`
-	Outcome        string `json:"outcome"`
-	EventTimestamp int64  `json:"eventTimestamp"`
+	ScriptName     string  `json:"scriptName"`
+	Outcome        string  `json:"outcome"`
+	EventTimestamp float64 `json:"eventTimestamp"`
 	Logs           []struct {
-		Message   []any  `json:"message"`
-		Level     string `json:"level"`
-		Timestamp int64  `json:"timestamp"`
+		Message   []any   `json:"message"`
+		Level     string  `json:"level"`
+		Timestamp float64 `json:"timestamp"`
 	} `json:"logs"`
 	Exceptions []struct {
-		Name      string `json:"name"`
-		Message   string `json:"message"`
-		Timestamp int64  `json:"timestamp"`
+		Name      string  `json:"name"`
+		Message   string  `json:"message"`
+		Timestamp float64 `json:"timestamp"`
 	} `json:"exceptions"`
 	Event struct {
 		Request struct {
@@ -79,13 +79,13 @@ func (c *cfCollector) Tail(ctx context.Context, opts Options, emit Emit) error {
 		return fmt.Errorf("cloudflare logs require --resource <worker-name>")
 	}
 	matcher := newMatcher(opts)
-	emitEntry := func(level, msg string, tsMs int64) error {
+	emitEntry := func(level, msg string, tsMs float64) error {
 		if strings.TrimSpace(msg) == "" {
 			return nil
 		}
 		ts := time.Now()
 		if tsMs > 0 {
-			ts = time.UnixMilli(tsMs)
+			ts = time.UnixMilli(int64(tsMs))
 		}
 		e := NewEntry("cloudflare", opts.Resource, opts.Resource, msg, ts)
 		if level != "" {
